+++ src/so2.rs
@@ -5,24 +5,24 @@ pub use crate::types::{Matrix2, Matrix3, Rotation2, Vector2, Vector3};
 #[allow(unused_imports)]
 use num_traits::real::Real;
 
-pub fn new_rotation2(theta: f64) -> Rotation2 {
+pub fn new_rotation2(theta: f32) -> Rotation2 {
     // In Rotation2::new is not supported in the no-std environment
     #[rustfmt::skip]
     Rotation2::from_matrix_unchecked(
         Matrix2::new(
-            f64::cos(theta), -f64::sin(theta),
-            f64::sin(theta), f64::cos(theta)
+            f32::cos(theta), -f32::sin(theta),
+            f32::sin(theta), f32::cos(theta)
         )
     )
 }
 
-pub fn log(rotation: &Matrix2) -> f64 {
-    f64::atan2(rotation[(1, 0)], rotation[(0, 0)])
+pub fn log(rotation: &Matrix2) -> f32 {
+    f32::atan2(rotation[(1, 0)], rotation[(0, 0)])
 }
 
-pub fn exp(theta: f64) -> Matrix2 {
-    let cos = f64::cos(theta);
-    let sin = f64::sin(theta);
+pub fn exp(theta: f32) -> Matrix2 {
+    let cos = f32::cos(theta);
+    let sin = f32::sin(theta);
     #[rustfmt::skip]
     Matrix2::new(
         cos, -sin,
@@ -33,7 +33,7 @@ pub fn exp(theta: f64) -> Matrix2 {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use core::f64::consts;
+    use core::f32::consts;
 
     #[test]
     fn test_exp() {
@@ -41,10 +41,10 @@ mod tests {
         let rot = exp(theta);
         assert_eq!(rot.nrows(), 2);
         assert_eq!(rot.ncols(), 2);
-        assert_eq!(rot[(0, 0)], f64::cos(theta));
-        assert_eq!(rot[(0, 1)], -f64::sin(theta));
-        assert_eq!(rot[(1, 0)], f64::sin(theta));
-        assert_eq!(rot[(1, 1)], f64::cos(theta));
+        assert_eq!(rot[(0, 0)], f32::cos(theta));
+        assert_eq!(rot[(0, 1)], -f32::sin(theta));
+        assert_eq!(rot[(1, 0)], f32::sin(theta));
+        assert_eq!(rot[(1, 1)], f32::cos(theta));
     }
 
     #[test]
