+++ src/types.rs
@@ -1,9 +1,9 @@
 use nalgebra::base::dimension::Const;
 use nalgebra::{ArrayStorage, Matrix, U1};
 
-pub type Vector<const D: usize> = Matrix<f64, Const<D>, U1, ArrayStorage<f64, D, 1>>;
-pub type MatrixNxN<const N: usize> = Matrix<f64, Const<N>, Const<N>, ArrayStorage<f64, N, N>>;
-pub type Rotation<const D: usize> = nalgebra::Rotation<f64, D>;
+pub type Vector<const D: usize> = Matrix<f32, Const<D>, U1, ArrayStorage<f32, D, 1>>;
+pub type MatrixNxN<const N: usize> = Matrix<f32, Const<N>, Const<N>, ArrayStorage<f32, N, N>>;
+pub type Rotation<const D: usize> = nalgebra::Rotation<f32, D>;
 
 pub type Matrix2 = MatrixNxN<2>;
 pub type Matrix3 = MatrixNxN<3>;
